Extracted method buildMatchFromToken on Lexer to simplify nextToken method. Added a few casts to better express conversion from int to other types specially ushort. Fixed typos on Lexer comments.

+++ DCPU16Emulator/Model/Emulator/Div.m
@@ -40,7 +40,7 @@
         [self.operationA setOverflowRegister:(ushort)((ushort)((divident << 16) / divisor) & 0xFFFF)];
     }
     
-    [self.operationA write:result];
+    [self.operationA write:(ushort) result];
     
     return result;
 }
